requirements.txt
alembic==1.12.0
attrs==21.4.0
bcrypt==3.2.2
blinker==1.6.2
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
default==0.1.2
Deprecated==1.2.13
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-SQLAlchemy-Session==1.1
Flask-Migrate==4.0.5
gevent
greenlet==2.0.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.25.2
packaging==21.3
pandas==1.4.3
pluggy==1.0.0
psycopg2-binary==2.9.7
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.27.1
six==1.16.0
SQLAlchemy==2.0.20
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.9
wrapt==1.13.3
Werkzeug==2.0.3
